refactor: 가로스크롤 자동 조정 모듈화
